pyproject.toml
[tool.poetry]
name = "photogalerybackend"
version = "0.1.0"
description = "Backend for PhotoGalery"
authors = ["Fred C <fcloitre@dbwildlife.info>"]
license = "AGPLv3"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
django-rest-framework = "^0.1.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
exif = "^1.6.1"
pillow = "^11.1.0"
psycopg = {extras = ["binary"], version = "^3.2.5"}
dj-database-url = "^2.3.0"
django-modeltranslation = "^0.19.12"
pre-commit = "^4.1.0"
djangorestframework-gis = "^1.1"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.0"
crispy-bootstrap5 = "^2024.10"
django-taggit = "^6.1.0"
sorl-thumbnail = "^12.11.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.5"
pylint-django = "^2.6.1"
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"

[tool.isort]
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
include = '\.pyi?$'
line-length = 79
target-version = ['py37']


[tool.flake8]
ignore = ['E231', 'E241', 'E501']
per-file-ignores = ['__init__.py:F401']
max-line-length = 79
count = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
